Добавил скрипт вставки тестовых значений Исправил ошибки с чат айди Исправил прикрепление питомца к усыновителю

+++ src/main/java/com/example/shelterbot/repository/CatsRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CatsRepository extends JpaRepository<Cats, Integer> {
+public interface CatsRepository extends JpaRepository<Cats, Long> {
 
 }
